update components and pages structure to render article and comments on new page, separating components
