+++ src/Components/Comments.jsx
@@ -1,14 +1,11 @@
-import { getCommentsByArticleId, getArticleById } from "../../../Utils/api";
-import Header from "../Header";
+import { getCommentsByArticleId, getArticleById } from "../../Utils/api";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import SingleArticle from "./SingleArticle";
-import CommentCard from "../CommentCard";
+import CommentCard from "../Components/CommentCard";
 
 const Comments = () => {
   const [comments, setComments] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
-  const [article, setArticle] = useState([]);
 
   const { articleId } = useParams();
 
@@ -19,21 +16,16 @@ const Comments = () => {
     });
   }, []);
 
-  useEffect(() => {
-    getArticleById(articleId).then((articleFromApi) => {
-      setArticle(articleFromApi);
-      setIsLoading(false);
-    });
-  }, []);
-
   if (isLoading) {
-    return <p>Loading...</p>;
+    return <p>Loading Comments...</p>;
+  }
+
+  if (comments.length === 0) {
+    return <p>It looks like nobody has commented on this article yet!</p>;
   }
 
   return (
     <section>
-      <Header title={`Comments on ${article.title}`} />
-      <SingleArticle />
       <ul className="comment-ul">
         {comments.map(({ author, body, created_at, votes, comment_id }) => {
           return (
